TwiM drift time calib fixed, suppress unnecessary cout
The drift time calibration for TwiM in s467/s444 (2020) has to be independent of the standard configuration since two MDPP16 were used for section 0 for those cases. The mapping of the time reference was wrong previously, and the R3BCoarseTimeStitch has been removed since it was causing fake reconstructions. Clang tidy for the TwimMapped2Cal and R3BCalifa applied to satisfy the requirements by the static-analysis in Github... New structures for handling Energy and Drift Time in Twim have been defined.