Updated IncomingBeta reconstruction for s091 experiment.
Clean-up and minor improvements in los classes.

Include timing of S2 hit after subtraction of the reference time in FrsSciPosCalData.

Bug fix in R3BFrsSciTcal2Cal when obtaining good pairs of hits of left and right PMTs.

Update IncomingData to reconstruct S2(FrsSci) to LosCal for s091 experiment. Los CFD out was used for the Tref signal for S2, so LosCal is used for reconstructing ToF specific to the experiment.
When multiple hits in Los are seen, only selecting good hit wrt the trigger timing is choosen.